added gme function for phylogenetics
