+++ skbio/tree/_nj.py
@@ -9,7 +9,6 @@
 import io
 
 import numpy as np
-import heapq as hq
 
 from skbio.stats.distance import DistanceMatrix
 from skbio.tree import TreeNode
@@ -103,6 +102,10 @@ def nj(dm, disallow_negative_branch_length=True, result_constructor=None):
     >>> newick_str = nj(dm, result_constructor=str)
     >>> print(newick_str[:55], "...")
     (d:2.000000, (c:4.000000, (b:3.000000, a:2.000000):3.00 ...
+
+    Notice that the tree constructed using neighbor joining is not rooted at a
+    leaf node, unlike minimum evolution, so re-rooting is required before
+    nearest neighbor interchange can be performed.
 
     """
     if dm.shape[0] < 3:
@@ -287,383 +290,3 @@ def _pair_members_to_new_node(dm, i, j, disallow_negative_branch_length):
         j_to_u = 0
 
     return i_to_u, j_to_u
-
-
-def nni(tree, dm, inplace=True):
-    r"""Perform nearest neighbor interchange (NNI) on a phylogenetic tree.
-
-    Parameters
-    ----------
-    tree : skbio.TreeNode
-        Input phylogenetic tree to be rearranged.
-    dm : skbio.DistanceMatrix
-        Input distance matrix containing distances between taxa.
-    inplace : bool, optional
-        Whether manipulate the tree in place (``True``, default) or return a
-        copy of the tree (``False``).
-
-    Returns
-    -------
-    TreeNode
-        Rearranged phylogenetic tree (if ``inplace`` is ``True``).
-
-    Notes
-    -----
-    NNI algorithm for minimum evolution problem on phylogenetic trees. It rearranges
-    an initial tree topology by performing subtree exchanges such that the distance
-    is minimized. This implementation is based on the FastNNI algorithm [1]_.
-
-    The input tree is required to be binary and rooted at a leaf node such that
-    there is a unique descendant from the root.
-
-    References
-    ----------
-    .. [1] Desper R, Gascuel O. Fast and accurate phylogeny reconstruction
-       algorithms based on the minimum-evolution principle. J Comput Biol.
-       2002;9(5):687-705. doi: 10.1089/106652702761034136. PMID: 12487758.
-
-    Examples
-    --------
-    Define a new distance matrix object describing the distances between five
-    taxa: human, monkey, pig, rat, and chicken.
-
-    >>> from skbio import DistanceMatrix
-    >>> from skbio.tree import nj
-
-    >>> dm = DistanceMatrix([[0, 0.02,  0.18,  0.34,  0.55],
-    ...                      [0.02,  0, 0.19, 0.35,  0.55],
-    ...                      [0.18, 0.19,  0,  0.34,  0.54],
-    ...                      [0.34, 0.35,  0.34,  0,  0.62],
-    ...                      [0.55,  0.55,  0.54,  0.62,  0]],
-    ...                      ['human','monkey','pig','rat','chicken'])
-
-    Also, provide a tree topology to be rearranged. The tree provided is
-    required to be a binary tree rooted at a leaf node.
-
-    Note that the tree provided does not require to have assigned edge lengths.
-
-    >>> from skbio.tree import TreeNode
-
-    >>> tree = TreeNode.read(["(((human,chicken),(rat,monkey)))pig;"])
-    >>> print(tree.ascii_art())
-                                  /-human
-                        /--------|
-                       |          \-chicken
-    -pig----- /--------|
-                       |          /-rat
-                        \--------|
-                                  \-monkey
-
-    Perform nearest neighbor interchange (NNI). By default, the tree is
-    rearrangede in place.
-
-    >>> nni(tree, dm)
-    >>> print(tree.ascii_art())
-                                  /-rat
-                        /--------|
-                       |          \-chicken
-    -pig----- /--------|
-                       |          /-monkey
-                        \--------|
-                                  \-human
-
-    Besides rearranging the tree, estimated edge lengths are assigned to the
-    tree.
-
-    >>> rat = tree.find('rat')
-    >>> print(rat.length)
-    0.21
-
-    """
-    # Initialize and populate the average distance matrix
-    if not inplace:
-        tree = tree.copy()
-    if len(tree.root().children) != 1:
-        raise TypeError(
-            "Could not perform NNI. " "Tree needs to be rooted at a leaf node."
-        )
-    for node in tree.non_tips():
-        if len(node.children) != 2:
-            raise TypeError("Could not perform NNI. Tree needs to be a binary tree.")
-    adm = _average_distance_matrix(tree, dm)
-    while True:
-        # create heap of possible swaps and then swapping subtrees
-        # until no more swaps are possible.
-        adm = _average_distance_matrix(tree, dm)
-        heap = _swap_heap(tree, adm)
-        if not heap:
-            break
-        swap = hq.heappop(heap)
-        _perform_swap(swap[1][0], swap[1][1])
-    # edge values are added using an OLS framework.
-    _edge_estimation(tree, dm)
-    if not inplace:
-        return tree
-
-
-def _perform_swap(node1, node2):
-    """Return a tree after swapping two subtrees."""
-    parent1, parent2 = node1.parent, node2.parent
-    parent1.append(node2)
-    parent2.append(node1)
-
-
-def _average_distance(node1, node2, dm):
-    """Return the average distance between the leaves of two subtrees.
-
-    Distances between nodes are calculated using a distance matrix.
-    """
-    nodelist1 = _tip_or_root(node1)
-    nodelist2 = _tip_or_root(node2)
-    df = dm.between(nodelist1, nodelist2)
-    return df["value"].mean()
-
-
-def _tip_or_root(node):
-    """Get name(s) of a node if it's a tip or root, otherwise its descending tips."""
-    if node.is_tip() or node.is_root():
-        return [node.name]
-    else:
-        return [x.name for x in node.tips()]
-
-
-def _average_distance_upper(node1, node2, dm):
-    """Return the average distance between the leaves of two subtrees.
-
-    Used for subtrees which have a set of tips that are the complement
-    of the set of tips that are descendants from the node defining
-    the subtree.
-
-    Given an internal edge of a binary tree, exactly one adjacent edge
-    will connect to a node defining a subtree of this form.
-    """
-    nodelist1 = _tip_or_root(node1)
-    if node2.is_root():
-        nodelist2 = []
-    # Second subtree serves as the tree with a set of tips
-    # complementary to the set of tips that descend from the
-    # corresponding second node.
-    else:
-        root2 = node2.root()
-        nodelist2 = [root2.name]
-        nodelist2.extend(root2.subset() - node2.subset())
-    df = dm.between(nodelist1, nodelist2)
-    return df["value"].mean()
-
-
-def _subtree_count(subtree):
-    """Return the number of leaves in a subtree.
-
-    Assumes the root as a leaf node.
-    """
-    if subtree.is_tip() or subtree.is_root():
-        return 1
-    else:
-        return subtree.count(tips=True)
-
-
-def _swap_length(a, b, c, d, i, j, k, m, adm):
-    """Return the change in overall tree length after a given swap.
-
-    The count of leaves contained in each subtree are denoted 'a, b, c, d' while
-    each node defining the subtree has the index 'i, j, k, m', respectively.
-    """
-    lambda1 = (a * d + b * c) / ((a + b) * (c + d))
-    lambda2 = (a * d + b * c) / ((a + c) * (b + d))
-    return 0.5 * (
-        (lambda1 - 1) * (adm[i][k] + adm[j][m])
-        - (lambda2 - 1) * (adm[i][j] + adm[k][m])
-        - (lambda1 - lambda2) * (adm[i][m] + adm[j][k])
-    )
-
-
-def _swap_heap(tree, adm):
-    """Return a maxheap ordered by the swap length for all possible swaps."""
-    heap = []
-    ordered = list(tree.postorder(include_self=False))
-    root = tree.root()
-    n_taxa = root.count(tips=True) + 1
-    # begin by finding nodes which are the child node of an internal edge
-    for node in ordered:
-        # ignore tips of the tree
-        if node.is_tip():
-            continue
-        # identify the parent and grandparent nodes
-        parent = node
-        a = parent.parent
-        # identify the index of each neighboring node
-        for index, node in enumerate(ordered):
-            if node == a:
-                i1 = index
-        for child in parent.children:
-            if child.is_tip():
-                continue
-            childnode = child
-            c, d = childnode.children
-            for sibling in childnode.siblings():
-                b = sibling
-            for index, node in enumerate(ordered):
-                if node == b:
-                    i2 = index
-                elif node == c:
-                    i3 = index
-                elif node == d:
-                    i4 = index
-            # count the tips of the subtrees defined by the neighboring nodes
-            sub_tips = []
-            for subtree in [b, c, d]:
-                sub_tips.append(1 if subtree.is_tip() else subtree.count(tips=True))
-            b_, c_, d_ = sub_tips
-            a_ = n_taxa - b_ - c_ - d_
-            # calculate the swap length for the two possible swaps given the edge
-            swap_1 = _swap_length(a_, b_, c_, d_, i1, i2, i3, i4, adm)
-            swap_2 = _swap_length(a_, b_, d_, c_, i1, i2, i4, i3, adm)
-            # store the best possible swap into a maxheap
-            if swap_1 > swap_2 and swap_1 > 0:
-                swap = -1 * swap_1
-                hq.heappush(heap, (swap, (b, c)))
-            elif swap_2 > swap_1 and swap_2 > 0:
-                swap = -1 * swap_2
-                hq.heappush(heap, (swap, (b, d)))
-    return heap
-
-
-def _average_subtree_distance(a, b, a1, a2, dm):
-    """Return the average distance between two subtrees."""
-    return (
-        _subtree_count(a1) * _average_distance(a1, b, dm)
-        + _subtree_count(a2) * _average_distance(a2, b, dm)
-    ) / _subtree_count(a)
-
-
-def _average_distance_matrix(tree, dm):
-    """Return the matrix of distances between pairs of subtrees."""
-    ordered = list(tree.postorder(include_self=False))
-    n = len(ordered)
-    r = tree.root()
-    taxa_size = r.count(tips=True) + 1
-    adm = np.empty((n, n))
-    for i, a in enumerate(ordered):
-        # skip over unique descendant
-        if a in tree.children:
-            continue
-        # find the average distance between given node and root
-        if a.is_tip():
-            adm[n - 1, i] = adm[i, n - 1] = dm[a.name, r.name]
-        else:
-            a1, a2 = a.children
-            adm[n - 1, i] = adm[i, n - 1] = _average_subtree_distance(a, r, a1, a2, dm)
-        # find the average distance between first node and a second node
-        # which is above the first node in the postorder as well as an ancestor
-        for j in range(i + 1, n - 1):  # part (a)
-            b = ordered[j]
-            # skipping over ancestors
-            if b in a.ancestors():
-                continue
-            # both nodes are tips
-            if a.is_tip() and b.is_tip():
-                adm[i, j] = adm[j, i] = dm[a.name, b.name]
-            # second node is a tip, but not the first node
-            elif b.is_tip():
-                a1, a2 = a.children
-                adm[i, j] = adm[j, i] = _average_subtree_distance(a, b, a1, a2, dm)
-            # neither node is a tip
-            else:
-                b1, b2 = b.children
-                adm[i, j] = adm[j, i] = _average_subtree_distance(b, a, b1, b2, dm)
-    # calculating for second nodes which are ancestors
-    for j, b in enumerate(ordered):
-        # skipping over unique descendant
-        if b in tree.children:
-            continue
-        s_ = b.siblings()
-        for sibling in s_:
-            s = sibling
-        p = b.parent
-        for i, a in enumerate(ordered):
-            if b in a.ancestors():
-                adm[i, j] = adm[j, i] = (
-                    _subtree_count(s) * _average_distance(a, s, dm)
-                    + (taxa_size - _subtree_count(p))
-                    * _average_distance_upper(a, p, dm)
-                ) / (taxa_size - _subtree_count(b))
-                # zero the diagonal
-                adm[i, i] = 0
-    return adm
-
-
-def _edge_estimation(tree, dm):
-    """Assign estimated edge values to a tree based on a given distance matrix.
-
-    Estimation of edge values is based on an ordinary least squares (OLS) framework.
-    """
-    adm = _average_distance_matrix(tree, dm)
-    ordered = list(tree.postorder(include_self=False))
-    root = tree.root()
-    taxa_size = root.count(tips=True) + 1
-    # identify edges by first finding the child node of an edge
-    for edge_node in ordered:
-        parent = edge_node.parent
-        # skip over root node
-        if edge_node.is_root():
-            continue
-        # calculate edge length for the edge adjacent to the root
-        elif parent.is_root():
-            for index, node in enumerate(ordered):
-                if node == edge_node:
-                    i1 = index
-            a, b = edge_node.children
-            for index, node in enumerate(ordered):
-                if node == a:
-                    i2 = index
-                elif node == b:
-                    i3 = index
-            edge_node.length = 0.5 * (adm[i2][i1] + adm[i3][i1] - adm[i2][i3])
-        # calculate edge lengths for external edges
-        elif edge_node.is_tip():
-            a = parent.parent
-            if a.is_root():
-                for child in a.children:
-                    a = child
-            for siblingnode in edge_node.siblings():
-                b = siblingnode
-                for index, node in enumerate(ordered):
-                    if node == edge_node:
-                        i1 = index
-                    if node == a:
-                        i2 = index
-                    if node == b:
-                        i3 = index
-            edge_node.length = 0.5 * (adm[i2][i1] + adm[i3][i1] - adm[i2][i3])
-        # calculate edge lengths for internal edges
-        else:
-            a = parent.parent
-            if a.is_root():
-                for child in a.children:
-                    a = child
-            for index, node in enumerate(ordered):
-                if node == a:
-                    i1 = index
-            c, d = edge_node.children
-            for sibling in edge_node.siblings():
-                b = sibling
-            for index, node in enumerate(ordered):
-                if node == b:
-                    i2 = index
-                elif node == c:
-                    i3 = index
-                elif node == d:
-                    i4 = index
-            # count the tips of subtrees which are adjacent to the internal edge
-            sub_tips = []
-            for subtree in [b, c, d]:
-                sub_tips.append(1 if subtree.is_tip() else subtree.count(tips=True))
-            b_, c_, d_ = sub_tips
-            a_ = taxa_size - b_ - c_ - d_
-            # calculate the edge length
-            lambda1 = (a_ * d_ + b_ * c_) / ((a_ + b_) * (c_ + d_))
-            edge_node.length = 0.5 * (
-                (lambda1 * (adm[i1][i3] + adm[i2][i4]))
-                + ((1 - lambda1) * (adm[i1][i4] + adm[i2][i3]))
-                - (adm[i1][i2] + adm[i3][i4])
-            )
